+++ maestro/src/ibis/maestro/Message.java
@@ -16,5 +16,5 @@ abstract class Message implements Serializable {
     /** Contractual obligation. */
     private static final long serialVersionUID = 1547379144090317151L;
 
-    long arrivalMoment;
+    double arrivalMoment;
 }
